LayoutTests/webcl/tests/webcl-test-arg-types-const-expected.txt

WebCL WebCLKernelArgumentTypes constants tests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test all constants defined in WebCLKernelArgumentTypes.idl
PASS WebCLKernelArgumentTypes.CHAR == 0 is true
PASS WebCLKernelArgumentTypes.UCHAR == 1 is true
PASS WebCLKernelArgumentTypes.SHORT == 2 is true
PASS WebCLKernelArgumentTypes.USHORT == 3 is true
PASS WebCLKernelArgumentTypes.INT == 4 is true
PASS WebCLKernelArgumentTypes.UINT == 5 is true
PASS WebCLKernelArgumentTypes.LONG == 6 is true
PASS WebCLKernelArgumentTypes.ULONG == 7 is true
PASS WebCLKernelArgumentTypes.FLOAT == 8 is true
PASS WebCLKernelArgumentTypes.HALF == 9 is true
PASS WebCLKernelArgumentTypes.DOUBLE == 10 is true
PASS WebCLKernelArgumentTypes.VEC2 == 0x0200 is true
PASS WebCLKernelArgumentTypes.VEC3 == 0x0300 is true
PASS WebCLKernelArgumentTypes.VEC4 == 0x0400 is true
PASS WebCLKernelArgumentTypes.VEC8 == 0x0800 is true
PASS WebCLKernelArgumentTypes.VEC16 == 0x1000 is true
PASS WebCLKernelArgumentTypes.LOCAL_MEMORY_SIZE == 255 is true
PASS successfullyParsed is true

TEST COMPLETE
